Split Milan feature extraction header
