+++ drivers/goodix53x5/milan/feature/extraction.c
@@ -8,7 +8,7 @@
  * version 2.1 of the License, or (at your option) any later version.
  */
 
-#include "milan/milan.h"
+#include "milan/feature/feature.h"
 #include "milan/private.h"
 
 #include <stddef.h>
